+++ resources/hpxml-measures/HPXMLtoOpenStudio/measure.rb
@@ -183,8 +183,11 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       end
 
       # Output
-      add_unmet_hours_output(model, hpxml_osm_map)
-      add_loads_output(model, args[:add_component_loads], hpxml_osm_map)
+      season_day_nums = add_unmet_hours_output(model, hpxml_osm_map)
+      loads_data = add_total_loads_output(model, hpxml_osm_map)
+      if args[:add_component_loads]
+        add_component_loads_output(model, hpxml_osm_map, loads_data, season_day_nums)
+      end
       set_output_files(model)
       add_additional_properties(model, hpxml, hpxml_osm_map, args[:hpxml_path], args[:building_id], hpxml_defaults_path)
       # Uncomment to debug EMS
@@ -459,7 +462,6 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     add_pools_and_permanent_spas(runner, model, spaces)
 
     # Other
-    add_cooling_season(model, weather)
     add_airflow(runner, model, weather, spaces, airloop_map)
     add_photovoltaics(model)
     add_generators(model)
@@ -1261,18 +1263,6 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     end
   end
 
-  def add_cooling_season(model, weather)
-    # Create cooling season schedule
-    # Applies to natural ventilation and calculation of component loads, not HVAC equipment
-    # Uses BAHSP cooling season, not user-specified cooling season (which may be, e.g., year-round)
-    _, default_cooling_months = HVAC.get_default_heating_and_cooling_seasons(weather, @hpxml_bldg.latitude)
-
-    clg_season_sch = MonthWeekdayWeekendSchedule.new(model, 'cooling season schedule', Array.new(24, 1), Array.new(24, 1), default_cooling_months, Constants.ScheduleTypeLimitsFraction)
-    @clg_ssn_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
-    @clg_ssn_sensor.setName('cool_season')
-    @clg_ssn_sensor.setKeyName(clg_season_sch.schedule.name.to_s)
-  end
-
   def add_windows(model, spaces)
     # We already stored @fraction_of_windows_operable, so lets remove the
     # fraction_operable properties from windows and re-collapse the enclosure
@@ -1935,7 +1925,7 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     end
 
     Airflow.apply(model, runner, weather, spaces, @hpxml_header, @hpxml_bldg, @cfa,
-                  @ncfl_ag, duct_systems, airloop_map, @clg_ssn_sensor, @eri_version,
+                  @ncfl_ag, duct_systems, airloop_map, @eri_version,
                   @frac_windows_operable, @apply_ashrae140_assumptions, @schedules_file,
                   @hpxml_header.unavailable_periods, hvac_availability_sensor)
   end
@@ -2082,7 +2072,7 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     # Create sensors and gather data
     htg_sensors, clg_sensors = {}, {}
     total_heat_load_serveds, total_cool_load_serveds = {}, {}
-    htg_start_days, htg_end_days, clg_start_days, clg_end_days = {}, {}, {}, {}
+    season_day_nums = {}
     hpxml_osm_map.each_with_index do |(hpxml_bldg, unit_model), unit|
       conditioned_zone_name = unit_model.getThermalZones.find { |z| z.additionalProperties.getFeatureAsString('ObjectType').to_s == HPXML::LocationConditionedSpace }.name.to_s
 
@@ -2102,10 +2092,12 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       next unless not hvac_control.nil?
 
       sim_year = @hpxml_header.sim_calendar_year
-      htg_start_days[unit] = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_begin_month, hvac_control.seasons_heating_begin_day)
-      htg_end_days[unit] = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_end_month, hvac_control.seasons_heating_end_day)
-      clg_start_days[unit] = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_begin_month, hvac_control.seasons_cooling_begin_day)
-      clg_end_days[unit] = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_end_month, hvac_control.seasons_cooling_end_day)
+      season_day_nums[unit] = {
+        htg_start: Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_begin_month, hvac_control.seasons_heating_begin_day),
+        htg_end: Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_end_month, hvac_control.seasons_heating_end_day),
+        clg_start: Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_begin_month, hvac_control.seasons_cooling_begin_day),
+        clg_end: Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_end_month, hvac_control.seasons_cooling_end_day)
+      }
     end
 
     hvac_availability_sensor = model.getEnergyManagementSystemSensors.find { |s| s.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameHVACAvailabilitySensor }
@@ -2120,10 +2112,10 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     program.addLine("Set #{clg_hrs} = 0")
     for unit in 0..hpxml_osm_map.size - 1
       if total_heat_load_serveds[unit] > 0
-        if htg_end_days[unit] >= htg_start_days[unit]
-          line = "If ((DayOfYear >= #{htg_start_days[unit]}) && (DayOfYear <= #{htg_end_days[unit]}))"
+        if season_day_nums[unit][:htg_end] >= season_day_nums[unit][:htg_start]
+          line = "If ((DayOfYear >= #{season_day_nums[unit][:htg_start]}) && (DayOfYear <= #{season_day_nums[unit][:htg_end]}))"
         else
-          line = "If ((DayOfYear >= #{htg_start_days[unit]}) || (DayOfYear <= #{htg_end_days[unit]}))"
+          line = "If ((DayOfYear >= #{season_day_nums[unit][:htg_start]}) || (DayOfYear <= #{season_day_nums[unit][:htg_end]}))"
         end
         line += " && (#{hvac_availability_sensor.name} == 1)" if not hvac_availability_sensor.nil?
         program.addLine(line)
@@ -2134,10 +2126,10 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       end
       next unless total_cool_load_serveds[unit] > 0
 
-      if clg_end_days[unit] >= clg_start_days[unit]
-        line = "If ((DayOfYear >= #{clg_start_days[unit]}) && (DayOfYear <= #{clg_end_days[unit]}))"
+      if season_day_nums[unit][:clg_end] >= season_day_nums[unit][:clg_start]
+        line = "If ((DayOfYear >= #{season_day_nums[unit][:clg_start]}) && (DayOfYear <= #{season_day_nums[unit][:clg_end]}))"
       else
-        line = "If ((DayOfYear >= #{clg_start_days[unit]}) || (DayOfYear <= #{clg_end_days[unit]}))"
+        line = "If ((DayOfYear >= #{season_day_nums[unit][:clg_start]}) || (DayOfYear <= #{season_day_nums[unit][:clg_end]}))"
       end
       line += " && (#{hvac_availability_sensor.name} == 1)" if not hvac_availability_sensor.nil?
       program.addLine(line)
@@ -2152,13 +2144,8 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     program_calling_manager.setName("#{program.name} calling manager")
     program_calling_manager.setCallingPoint('EndOfZoneTimestepBeforeZoneReporting')
     program_calling_manager.addProgram(program)
-  end
 
-  def add_loads_output(model, add_component_loads, hpxml_osm_map)
-    loads_data = add_total_loads_output(model, hpxml_osm_map)
-    return unless add_component_loads
-
-    add_component_loads_output(model, hpxml_osm_map, loads_data)
+    return season_day_nums
   end
 
   def add_total_loads_output(model, hpxml_osm_map)
@@ -2246,7 +2233,7 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     return htg_cond_load_sensors, clg_cond_load_sensors, total_heat_load_serveds, total_cool_load_serveds, dehumidifier_sensors
   end
 
-  def add_component_loads_output(model, hpxml_osm_map, loads_data)
+  def add_component_loads_output(model, hpxml_osm_map, loads_data, season_day_nums)
     htg_cond_load_sensors, clg_cond_load_sensors, total_heat_load_serveds, total_cool_load_serveds, dehumidifier_sensors = loads_data
 
     # Output diagnostics needed for some output variables used below
@@ -2599,6 +2586,23 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
         program.addLine("Set hr_ducts = hr_ducts + (#{ducts_mix_loss_sensor.name} - #{ducts_mix_gain_sensor.name})")
       end
 
+      # EMS Sensors: Indoor temperature, setpoints
+      tin_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Zone Mean Air Temperature')
+      tin_sensor.setName('tin s')
+      tin_sensor.setKeyName(conditioned_zone.name.to_s)
+      thermostat = nil
+      if conditioned_zone.thermostatSetpointDualSetpoint.is_initialized
+        thermostat = conditioned_zone.thermostatSetpointDualSetpoint.get
+
+        htg_sp_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
+        htg_sp_sensor.setName('htg sp s')
+        htg_sp_sensor.setKeyName(thermostat.heatingSetpointTemperatureSchedule.get.name.to_s)
+
+        clg_sp_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
+        clg_sp_sensor.setName('clg sp s')
+        clg_sp_sensor.setKeyName(thermostat.coolingSetpointTemperatureSchedule.get.name.to_s)
+      end
+
       # EMS program: Heating vs Cooling logic
       program.addLine('Set htg_mode = 0')
       program.addLine('Set clg_mode = 0')
@@ -2606,10 +2610,40 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       program.addLine("  Set htg_mode = #{total_heat_load_serveds[unit]}")
       program.addLine("ElseIf (#{clg_cond_load_sensors[unit].name} > 0)") # Assign hour to cooling if cooling load
       program.addLine("  Set clg_mode = #{total_cool_load_serveds[unit]}")
-      program.addLine("ElseIf (#{@clg_ssn_sensor.name} > 0)") # No load, assign hour to cooling if in cooling season definition (Note: natural ventilation & whole house fan only operate during the cooling season)
-      program.addLine("  Set clg_mode = #{total_cool_load_serveds[unit]}")
-      program.addLine('Else') # No load, assign hour to heating if not in cooling season definition
-      program.addLine("  Set htg_mode = #{total_heat_load_serveds[unit]}")
+      program.addLine('Else')
+      program.addLine('  Set htg_season = 0')
+      program.addLine('  Set clg_season = 0')
+      if not season_day_nums[unit].nil?
+        # Determine whether we're in the heating and/or cooling season
+        if season_day_nums[unit][:clg_end] >= season_day_nums[unit][:clg_start]
+          program.addLine("  If ((DayOfYear >= #{season_day_nums[unit][:clg_start]}) && (DayOfYear <= #{season_day_nums[unit][:clg_end]}))")
+        else
+          program.addLine("  If ((DayOfYear >= #{season_day_nums[unit][:clg_start]}) || (DayOfYear <= #{season_day_nums[unit][:clg_end]}))")
+        end
+        program.addLine('    Set clg_season = 1')
+        program.addLine('  EndIf')
+        if season_day_nums[unit][:htg_end] >= season_day_nums[unit][:htg_start]
+          program.addLine("  If ((DayOfYear >= #{season_day_nums[unit][:htg_start]}) && (DayOfYear <= #{season_day_nums[unit][:htg_end]}))")
+        else
+          program.addLine("  If ((DayOfYear >= #{season_day_nums[unit][:htg_start]}) || (DayOfYear <= #{season_day_nums[unit][:htg_end]}))")
+        end
+        program.addLine('    Set htg_season = 1')
+        program.addLine('  EndIf')
+      end
+      program.addLine("  If ((#{natvent_sensors[0].name} <> 0) || (#{natvent_sensors[1].name} <> 0)) && (clg_season == 1)") # Assign hour to cooling if natural ventilation is operating
+      program.addLine("    Set clg_mode = #{total_cool_load_serveds[unit]}")
+      program.addLine("  ElseIf ((#{whf_sensors[0].name} <> 0) || (#{whf_sensors[1].name} <> 0)) && (clg_season == 1)") # Assign hour to cooling if whole house fan is operating
+      program.addLine("    Set clg_mode = #{total_cool_load_serveds[unit]}")
+      if not thermostat.nil?
+        program.addLine('  Else') # Indoor temperature floating between setpoints; determine assignment by comparing to average of heating/cooling setpoints
+        program.addLine("    Set Tmid_setpoint = (#{htg_sp_sensor.name} + #{clg_sp_sensor.name}) / 2")
+        program.addLine("    If (#{tin_sensor.name} > Tmid_setpoint) && (clg_season == 1)")
+        program.addLine("      Set clg_mode = #{total_cool_load_serveds[unit]}")
+        program.addLine("    ElseIf (#{tin_sensor.name} < Tmid_setpoint) && (htg_season == 1)")
+        program.addLine("      Set htg_mode = #{total_heat_load_serveds[unit]}")
+        program.addLine('    EndIf')
+      end
+      program.addLine('  EndIf')
       program.addLine('EndIf')
 
       unit_multiplier = @hpxml_bldg.building_construction.number_of_units