+++ resources/hpxml-measures/ReportSimulationOutput/measure.rb
@@ -2871,7 +2871,6 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
           Constants.ObjectNameMiscPermanentSpaHeater => EUT::PermanentSpaHeater,
           Constants.ObjectNameMechanicalVentilationPreheating => EUT::MechVentPreheat,
           Constants.ObjectNameMechanicalVentilationPrecooling => EUT::MechVentPrecool,
-          Constants.ObjectNameWaterHeaterAdjustment => EUT::HotWater,
           Constants.ObjectNameBatteryLossesAdjustment => EUT::Battery }.each do |obj_name, eut|
           next unless subcategory.start_with? obj_name
           fail 'Unepected error: multiple matches.' unless end_use.nil?